fix: 로또 구입 금액 형식 int -> long 수정

+++ src/main/java/lotto/controller/LottoGameController.java
@@ -28,7 +28,7 @@ public class LottoGameController {
     private void prepareMoney() {
         OutputView.printPurchasePrice();
         try {
-            int lottoPurchase = InputView.inputLottoPurchase();
+            long lottoPurchase = InputView.inputLottoPurchase();
             lottoCount = LottoCount.calculatePurchaseCount(lottoPurchase);
         } catch (IllegalArgumentException e) {
             OutputView.printError(e);
